+++ extensions/cornerstone/src/types/WindowLevel.ts
@@ -1,0 +1,5 @@
+export type WindowLevelPreset = {
+  description: string;
+  window: string;
+  level: string;
+};
